+++ modules/kubernetes/pvc/outputs.tf
@@ -1,0 +1,3 @@
+output "name" {
+  value = kubernetes_persistent_volume_claim_v1.volume.metadata.0.name
+}